+++ docs/CHANGELOG.rst
@@ -13,6 +13,8 @@ Unreleased
 Added
 -----
 - Method ``groups_with_permissions`` was added to the ``PermissionObject``
+- Register annotation models with Django admin
+- Enable editing ``AnnotationPreset`` permissions in the Django admin
 
 Changed
 -------